feat(auth): 添加验证码生成与校验功能
新增验证码相关的工具函数，支持生成和校验 base64使用 验证码， Redis 存储验证码信息。同时在登录接口中增加验证码校验逻辑，并更新了相关错误码定义。

- 在 `internal/util` 包中新增 `captcha_generate.go` 文件，实现 `GetCaptcha` 和 `VerifyCaptcha` 函数
- 更新 `internal/common/errorcode.go`，添加验证码相关错误码- 在 `configs/config.yaml` 中排除验证码接口路径
- 调整 `GetUserIdFromContext` 返回值类型为 int64- 移除 `internal/module/common_filed.go` 中的通用字段定义- 引入 `base64Captcha` 和 swagger 相关依赖包
- 添加数据库代码生成脚本 `cmd/generate/generate.go`

+++ internal/util/context_utils.go
@@ -17,6 +17,6 @@ func GetUserInfoFromContext(c *gin.Context) *Claims {
 }
 
 // GetUserIdFromContext 从上下文中获取用户ID
-func GetUserIdFromContext(c *gin.Context) uint {
+func GetUserIdFromContext(c *gin.Context) int64 {
 	return GetUserInfoFromContext(c).GetUserId()
 }
